Move code to MapGen, add AgentsGen
 #7
